+++ FireWithin/history.md
@@ -27,6 +27,7 @@ on this page.
 
 Release           | What                 | *When*
 ------------------|----------------------|------------------
+[v24.0](#v24_0)   | Memorize God's Word & more              | *Jan 29, 2025*
 [v23.0](#v23_0)   | Questions and Answers & more            | *Jan 7, 2025*
 [v22.7](#v22_7)   | Update Current Sermon Series            | *Jan 1, 2025*
 [v22.6](#v22_6)   | Update Current Sermon Series            | *Dec 3, 2024*
@@ -67,6 +68,21 @@ Release           | What                 | *When*
 
 ## Details:
 
+
+<!-- ************************************************************* -->
+<br/>
+<h3 id="v24_0" style="margin: 10px 0px; border-width: 5px 0px; padding: 5px; border-style: solid;">
+  v24.0 - Memorize God's Word & more <i>(Jan 29, 2025)</i>
+</h3>
+<br/>
+
+- Added new {{book.Memorization}} page.
+- Update Current {{book.2Corinthians}} Series (Sundays)
+- Update Current {{book.1Kings}} Series (Mid Week)
+- Added 2025 entries under {{book.SpecialEvents}} / {{book.CurrentEvents}} for:
+  * {{book.Life}}
+  * {{book.ElectionDaySermons}}
+
 <!-- ************************************************************* -->
 <br/>
 <h3 id="v23_0" style="margin: 10px 0px; border-width: 5px 0px; padding: 5px; border-style: solid;">
@@ -76,6 +92,7 @@ Release           | What                 | *When*
 
 - Added new {{book.QA}} section.
 - Update Current {{book.2Corinthians}} Series (Sundays)
+- Added {{book.EyeWitness}} reference to the {{book.Acts}} Series
 
 
 <!-- ************************************************************* -->